Split up snapshot class into modules
